requirements.txt
Flask
docker
python-dotenv
flask-cors
Flask-HTTPAuth